chore(provider): remove legacy adapter aggregate
