+++ src/application/services/provider-adapters/index.ts
@@ -1,0 +1,5 @@
+export { ProviderAdapter } from './provider-adapter.js';
+export { OllamaAdapter } from './ollama-adapter.js';
+export { LMStudioAdapter } from './lm-studio-adapter.js';
+export { ClaudeAdapter } from './claude-adapter.js';
+export { HuggingFaceAdapter } from './huggingface-adapter.js';
